+++ src/reader/app/types.ts
@@ -299,6 +299,7 @@ export interface ReaderSettings {
     immersionKitPlayOnHover: boolean;
     immersionKitPlayOnImageClick: boolean;
     immersionKitPlaybackRate: number;
+    selectionPopoverShowTranslation: boolean;
     parseSelection: boolean;
     lookupOnClick: boolean;
     lookupOnHover: boolean;